requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
openai
langchain
joblib
tqdm
